Bokshare Application 2017
The purpose of this project is to allow particulars to be able to buy textbooks at a lower price due to the excessive price of textbooks that many students are complaining about. Particulars will also benefit by selling the textbooks they are no longer using. Our platform aims to bring people together (Seller and Buyer).